// src/enums/Roles.ts
export enum TypeRoles {
  ALL = 'All',
  ANALYST = 'Analyst',
  KADIV = 'KADIV',
  RM = 'STAFF',
  SUPER_ADMIN = 'Super Admin',
  TL = 'TL',
  TL_ANALYST = 'TL Analyst',
  STAFF_FM = 'Staff_FM',
  FM = 'FM',
  STAFF_RM = 'RM',
  ARM = 'ARM',
  STAFF_ANALYST = 'STAFF_ANALYST',
  TECHNICAL_STAFF = 'TECHNICAL_STAFF',
  GENERAL_STAFF = 'GENERAL_STAFF'
};